requirements.txt
asgiref==3.11.1
Django==6.0.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg==3.3.3
psycopg2==2.9.11
PyJWT==2.11.0
redis==7.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
